refactor(burger-ingredients): moved the ingredients request to burger-ingredients
